apps/image-processor/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# AI-powered background removal
rembg==2.0.50
torch==2.0.1
torchvision==0.15.2

# Real-ESRGAN and upscaling
basicsr==1.4.2
realesrgan==0.3.0
gfpgan==1.3.8

# SVG processing
potrace==0.1.0
svgwrite==1.4.3
lxml==4.9.3

# Database and storage
supabase==2.0.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
